// API/src/shared/container/providers/MailProvider/implementations/EtherealMailProvider.ts
import nodemailer, { Transporter } from 'nodemailer';
import { inject, injectable } from 'tsyringe';
import IMailTemplateProvider from '@shared/container/providers/MailTemplateProvider/models/IMailTemplateProvider';
import IMailProvider from '../models/IMailProvider';
import ISendMailDTO from '../dtos/ISendMailDTO';

@injectable()
class EtherealMailProvider implements IMailProvider {
	private client: Transporter;

	constructor(
		@inject('MailTemplateProvider')
		private mailTemplateProvider: IMailTemplateProvider,
	) {
		const transporter = nodemailer.createTransport({
			host: 'smtp.ethereal.email',
			port: 587,
			auth: {
				user: 'leta3@ethereal.email',
				pass: 'cy8KGAygMYZXyB7hH4',
			},
		});
		this.client = transporter;
	}

	public async sendMail({
		from,
		to,
		subject,
		templateData,
	}: ISendMailDTO): Promise<void> {
		await this.client.sendMail({
			from: {
				name: from?.name || 'Equipe Gobarber',
				address: from?.email || 'suporte@gobarber.com.br',
			},
			to: {
				name: to.name,
				address: to.email,
			},
			subject,
			html: await this.mailTemplateProvider.parser(templateData),
		});
	}
}

export default EtherealMailProvider;
